+++ src/components/settings/language.tsx
@@ -3,17 +3,24 @@ import { Field, reduxForm, WrappedFieldProps } from "redux-form";
 import Header from "semantic-ui-react/dist/es/elements/Header";
 import Dropdown from "semantic-ui-react/dist/es/modules/Dropdown";
 
+import DefaultLanguageInput from "./defaultlanguageinput";
 import LanguageDropdownInput from "./languagedropdowninput";
 
 
 const LanguageSettings = (props: {}) => (
   <div>
     <Header dividing={true}>
-      Highlight Language Options
+      Highlight Languages
     </Header>
     <Field
       name="languages"
       component={LanguageDropdownInput}
+      style={{ marginBottom: "1em" }}
+    />
+
+    <Field
+      name="defaultlanguage"
+      component={DefaultLanguageInput as any}
     />
   </div>
 );